+++ login/login_helper.go
@@ -1,97 +1,28 @@
 package login
 
 import (
-	"bufio"
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/md5"
-	"crypto/rand"
 	"encoding/base64"
-	"encoding/hex"
-	"errors"
-	"fmt"
-	"io"
-	"log"
-	mathrand "math/rand"
-	"os"
-	"path/filepath"
 	"regexp"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/jinzhu/gorm"
-	"github.com/pass-wall/passwall-server/pkg/database"
+	"github.com/pass-wall/passwall-server/helper"
 	"github.com/spf13/viper"
 )
 
-// AddValues ...
-func AddValues(url, username, password string, file *os.File) error {
-	db := database.GetDB()
-	var urlIndex, usernameIndex, passwordIndex int
-
-	scanner := bufio.NewScanner(file)
-	counter := 0
-	for scanner.Scan() {
-		fields := strings.Split(scanner.Text(), ",")
-
-		// Ignore first line (field names)
-		counter++
-		if counter == 1 {
-			// Match user's fieldnames to passwall's field names (URL, Username, Password)
-			urlIndex = FindIndex(fields, url)
-			usernameIndex = FindIndex(fields, username)
-			passwordIndex = FindIndex(fields, password)
-
-			// Check if fields match
-			if urlIndex == -1 || usernameIndex == -1 || passwordIndex == -1 {
-				errorText := fmt.Sprintf("%s, %s or %s field couldn't found in %s file", url, username, password, filepath.Base(file.Name()))
-				err := errors.New(errorText)
-				return err
-			}
+// DecryptLoginPasswords ...
+func DecryptLoginPasswords(logins []Login) []Login {
+	for i := range logins {
+		if logins[i].Password == "" {
 			continue
 		}
-
-		// if isRecordNotFound(fields[urlIndex], fields[usernameIndex], fields[passwordIndex]) {
-		// Fill login struct with csv file content
-		login := Login{
-			URL:      fields[urlIndex],
-			Username: fields[usernameIndex],
-			Password: base64.StdEncoding.EncodeToString(Encrypt(fields[passwordIndex], viper.GetString("server.passphrase"))),
-		}
-
-		// Add to database
-		db.Create(&login)
+		passByte, _ := base64.StdEncoding.DecodeString(logins[i].Password)
+		passB64 := string(helper.Decrypt(string(passByte[:]), viper.GetString("server.passphrase")))
+		logins[i].Password = passB64
 	}
-
-	if err := scanner.Err(); err != nil {
-		log.Println(err)
-		return err
-	}
-
-	return nil
-}
-
-func FindIndex(vs []string, t string) int {
-	for i, v := range vs {
-		if v == t {
-			return i
-		}
-	}
-	return -1
-}
-
-// CheckErr ...
-func CheckErr(err error) {
-	if err != nil {
-		log.Println(err)
-	}
-}
-
-// Include ...
-func Include(vs []string, t string) bool {
-	return FindIndex(vs, t) >= 0
+	return logins
 }
 
 // SetArgs ...
@@ -152,7 +83,7 @@ func setOrder(sort, order string) string {
 	sortValues := []string{"id", "created_at", "updated_at", "url", "username"}
 	orderValues := []string{"desc", "asc"}
 
-	if Include(sortValues, ToSnakeCase(sort)) && Include(orderValues, ToSnakeCase(order)) {
+	if helper.Include(sortValues, ToSnakeCase(sort)) && helper.Include(orderValues, ToSnakeCase(order)) {
 		return ToSnakeCase(sort) + " " + ToSnakeCase(order)
 	}
 
@@ -187,76 +118,4 @@ func ToSnakeCase(str string) string {
 	snake = matchAllCap.ReplaceAllString(snake, "${1}_${2}")
 
 	return strings.ToLower(snake)
-}
-
-// Password ..
-func Password() string {
-	mathrand.Seed(time.Now().UnixNano())
-	chars := []rune("ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
-		"abcdefghijklmnopqrstuvwxyz" +
-		"0123456789" +
-		"=+%*/()[]{}/!@#$?|")
-	length := viper.GetInt("server.generatedPasswordLength")
-	var b strings.Builder
-	for i := 0; i < length; i++ {
-		b.WriteRune(chars[mathrand.Intn(len(chars))])
-	}
-	return b.String()
-}
-
-// CreateHash ...
-func CreateHash(key string) string {
-	hasher := md5.New()
-	hasher.Write([]byte(key))
-	return hex.EncodeToString(hasher.Sum(nil))
-}
-
-// Encrypt ..
-func Encrypt(dataStr string, passphrase string) []byte {
-	dataByte := []byte(dataStr)
-	block, _ := aes.NewCipher([]byte(CreateHash(passphrase)))
-	gcm, err := cipher.NewGCM(block)
-	if err != nil {
-		panic(err.Error())
-	}
-	nonce := make([]byte, gcm.NonceSize())
-	if _, err = io.ReadFull(rand.Reader, nonce); err != nil {
-		panic(err.Error())
-	}
-	cipherByte := gcm.Seal(nonce, nonce, dataByte, nil)
-	return cipherByte
-}
-
-// Decrypt ...
-func Decrypt(dataStr string, passphrase string) string {
-	dataByte := []byte(dataStr)
-	key := []byte(CreateHash(passphrase))
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		panic(err.Error())
-	}
-	gcm, err := cipher.NewGCM(block)
-	if err != nil {
-		panic(err.Error())
-	}
-	nonceSize := gcm.NonceSize()
-	nonce, ciphertext := dataByte[:nonceSize], dataByte[nonceSize:]
-	plainByte, err := gcm.Open(nil, nonce, ciphertext, nil)
-	if err != nil {
-		panic(err.Error())
-	}
-	return string(plainByte[:])
-}
-
-// DecryptLoginPasswords ...
-func DecryptLoginPasswords(logins []Login) []Login {
-	for i := range logins {
-		if logins[i].Password == "" {
-			continue
-		}
-		passByte, _ := base64.StdEncoding.DecodeString(logins[i].Password)
-		passB64 := Decrypt(string(passByte[:]), viper.GetString("server.passphrase"))
-		logins[i].Password = passB64
-	}
-	return logins
 }