+++ pkg/router/router.go
@@ -1,15 +1,14 @@
 package router
 
 import (
-	"log"
-
-	jwt "github.com/appleboy/gin-jwt/v2"
+	"github.com/gin-contrib/secure"
+	"github.com/gin-contrib/static"
+	"github.com/gin-gonic/gin"
 	"github.com/jinzhu/gorm"
 	"github.com/pass-wall/passwall-server/login"
 	"github.com/pass-wall/passwall-server/pkg/database"
 	"github.com/pass-wall/passwall-server/pkg/middleware"
-
-	"github.com/gin-gonic/gin"
+	"github.com/pass-wall/passwall-server/util"
 )
 
 // Setup initializes the gin engine and router
@@ -20,6 +19,10 @@ func Setup() *gin.Engine {
 	r.Use(gin.Logger())
 	r.Use(gin.Recovery())
 	r.Use(middleware.CORS())
+	r.Use(secure.New(secureConfig()))
+
+	// Serve static files in public folder
+	r.Use(static.Serve("/", static.LocalFile("./public", true)))
 
 	db := database.GetDB()
 	loginAPI := InitLoginAPI(db)
@@ -29,7 +32,7 @@ func Setup() *gin.Engine {
 
 	auth := r.Group("/auth")
 	{
-		auth.POST("/signin", authMW.LoginHandler)
+		auth.POST("/signin", middleware.LimiterMW(), authMW.LoginHandler)
 		auth.POST("/check", authMW.MiddlewareFunc(), middleware.TokenCheck)
 		auth.POST("/refresh", authMW.MiddlewareFunc(), authMW.RefreshHandler)
 	}
@@ -40,16 +43,22 @@ func Setup() *gin.Engine {
 		logins.GET("/", loginAPI.FindAll)
 		logins.GET("/:id", loginAPI.FindByID)
 		logins.POST("/", loginAPI.Create)
-		logins.POST("/:action", login.PostHandler)
+		logins.POST("/:action", func(c *gin.Context) {
+			path := c.Param("action")
+			if path == "check-password" {
+				loginAPI.FindSamePassword(c)
+			} else {
+				util.PostHandler(c)
+			}
+		})
+
 		logins.PUT("/:id", loginAPI.Update)
 		logins.DELETE("/:id", loginAPI.Delete)
+
 	}
 
-	// Protection for route/endpoint scaning
-	r.NoRoute(authMW.MiddlewareFunc(), func(c *gin.Context) {
-		claims := jwt.ExtractClaims(c)
-		log.Printf("NoRoute claims: %#v\n", claims)
-		c.JSON(404, gin.H{"status": "Error", "message": "Page not found"})
+	r.NoRoute(func(c *gin.Context) {
+		c.File("./public/index.html")
 	})
 
 	return r
@@ -62,4 +71,23 @@ func InitLoginAPI(db *gorm.DB) login.LoginAPI {
 	loginAPI := login.NewLoginAPI(loginService)
 	loginAPI.Migrate()
 	return loginAPI
+}
+
+func secureConfig() secure.Config {
+	// Details about this config is here
+	// https://github.com/gin-contrib/secure/blob/master/secure.go
+	return secure.Config{
+		// AllowedHosts:          []string{"example.com", "ssl.example.com"},
+		// SSLRedirect:           false,
+		// SSLHost:               "ssl.example.com",
+		STSSeconds:            315360000,
+		STSIncludeSubdomains:  true,
+		FrameDeny:             true,
+		ContentTypeNosniff:    true,
+		BrowserXssFilter:      true,
+		ContentSecurityPolicy: "default-src 'self' 'unsafe-inline' 'unsafe-eval'; connect-src *",
+		IENoOpen:              true,
+		ReferrerPolicy:        "strict-origin-when-cross-origin",
+		SSLProxyHeaders:       map[string]string{"X-Forwarded-Proto": "https"},
+	}
 }